requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
